Only (re-)enable interrupts on FIFO_NEED_MORE
With the current logic, it is the only state where we can be sure that
FIFO_STATUS_AF has an edge that triggers EVENT_AF when there is new
data. This is not a problem, and so there is no reason to enable an
interrupt that might or might not occur in other states.