Quartus: Interrupts on fifo_h2f
Furthermore: exchanged the FIFO depths for the two connected FIFO's in
the path from HPS to FPGA. Previously, the FIFO on the FPGA side was 512
32-bit words deep, and the FIFO on the HPS side was 8 words deep.

The new situation means that if the FPGA falls behind on reading the
FIFO, it will read the level register once for every 8 32-bit words of
data. Note that the status register is slower than the data register.
This limits the maximum speed in theory, but likely has no practical
effect. On the other hand, the HPS now sees a 512-word deep FIFO, which
lowers the interrupt pressure on the HPS when the FPGA doesn't
immediately empty the FIFO.